backend/requirements.txt
flask==3.0.3
flask-cors==5.0.0
python-dotenv==1.0.1
google-generativeai==0.8.3

